requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.11.7
    # via discord-py
aiosignal==1.3.1
    # via aiohttp
attrs==24.2.0
    # via aiohttp
build==1.2.2.post1
    # via pip-tools
click==8.1.7
    # via pip-tools
discord-py==2.4.0
    # via -r requirements.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.10
    # via yarl
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
packaging==24.2
    # via build
pip-tools==7.4.1
    # via -r requirements.in
propcache==0.2.0
    # via
    #   aiohttp
    #   yarl
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dotenv==1.0.1
    # via -r requirements.in
ruff==0.6.9
    # via -r requirements.in
wheel==0.45.1
    # via pip-tools
yarl==1.18.0
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
